requirements.txt
celery==5.2.3
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
redis==4.1.1